# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Vecino Vigilante - API de contrataciones públicas de Chanchamayo y Junín
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /stats:
    get:
      operationId: getStats
      tags: [stats]
      summary: KPIs globales del portal
      parameters:
        - name: anio
          in: query
          schema:
            type: ["integer", "null"]
        - name: ubigeo
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Estadísticas globales
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatsGlobales"
  /stats/evolucion:
    get:
      operationId: getEvolucionMensual
      tags: [stats]
      summary: Evolución mensual de contrataciones
      parameters:
        - name: ubigeo
          in: query
          schema:
            type: ["string", "null"]
        - name: anios
          in: query
          schema:
            type: ["integer", "null"]
            default: 2
      responses:
        "200":
          description: Serie mensual
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PuntoEvolucion"
  /stats/tipo-distribucion:
    get:
      operationId: getTipoDistribucion
      tags: [stats]
      summary: Distribución de contrataciones por tipo (obras, bienes, servicios, consultoría)
      parameters:
        - name: ubigeo
          in: query
          schema:
            type: ["string", "null"]
        - name: anio
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Distribución por tipo
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TipoDistribucion"
...
components:
  schemas:
...
    StatsGlobales:
      type: object
      properties:
        totalContrataciones:
          type: integer
        montoTotal:
          type: ["number", "null"]
        contratacionesEsteAnio:
          type: integer
        montoEsteAnio:
          type: ["number", "null"]
        obrasEnEjecucion:
          type: integer
        entidadesActivas:
          type: integer
        proveedoresUnicos:
          type: integer
        ultimaSincronizacion:
          type: ["string", "null"]
        porcentajeDirectas:
          type: ["number", "null"]
        porcentajeCompetitivas:
          type: ["number", "null"]
      required:
        - totalContrataciones
        - contratacionesEsteAnio
        - obrasEnEjecucion
        - entidadesActivas
        - proveedoresUnicos
    PuntoEvolucion:
      type: object
      properties:
        mes:
          type: string
        anio:
          type: integer
        periodo:
          type: string
        totalContrataciones:
          type: integer
        montoTotal:
          type: ["number", "null"]
      required:
        - mes
        - anio
        - periodo
        - totalContrataciones
    TipoDistribucion:
      type: object
      properties:
        tipo:
          type: string
        total:
          type: integer
        monto:
          type: ["number", "null"]
        porcentaje:
          type: number
      required:
        - tipo
        - total
        - porcentaje
...